Updated models.py.

+++ models.py
@@ -4,10 +4,10 @@ def create_classes(db):
 
         id = db.Column(db.Integer, primary_key=True)
         expenditures = db.Column(db.String(255))
-        actual_fy_18_19 = db.Column(db.Numeric)
-        adopted_fy_19_20 = db.Column(db.Numeric)
-        estimated_fy_19_20 = db.Column(db.Numeric)
-        adopted_fy_20_21 = db.Column(db.Numeric)
+        actual_fy_18_19 = db.Column(db.Float)
+        adopted_fy_19_20 = db.Column(db.Float)
+        estimated_fy_19_20 = db.Column(db.Float)
+        adopted_fy_20_21 = db.Column(db.Float)
 
         def __repr__(self):
             return '<Budget %r>' % (self.name)
